`qvf-cli` utility endpoint for constructor indices
