+++ quadraticVoting/quadraticVoting.cabal
@@ -20,7 +20,8 @@ library
 
   default-language:    Haskell2010
 
-  exposed-modules:     QVF
+  exposed-modules:     CLI.Utils
+                     , QVF
                      , Data.Datum
                      , Data.DonationInfo
                      , Data.Redeemer
@@ -42,6 +43,8 @@ library
                      , plutus-tx-plugin
                      , plutus-tx
                      , serialise
+                     , text
+                     , time
 
   ghc-options:         -Wall 
                        -fobject-code
